Make the new QNode default
