+++ pennylane/__init__.py
@@ -50,6 +50,7 @@ from pennylane.templates.subroutines import *
 from pennylane.optimize import *
 from pennylane import qaoa
 from pennylane.qnode import QNode, qnode
+import pennylane.qnode_old
 from pennylane.transforms import (
     adjoint,
     batch_params,
