+++ pennylane/transforms/batch_transform.py
@@ -20,7 +20,6 @@ import types
 import warnings
 
 import pennylane as qml
-from pennylane.beta import QNode
 
 
 class batch_transform:
@@ -325,7 +324,7 @@ class batch_transform:
             # tapes, fn = some_transform(tape, *transform_args)
             return self.construct(qnode, *targs, **tkwargs)
 
-        if isinstance(qnode, (qml.QNode, QNode, qml.ExpvalCost)):
+        if isinstance(qnode, (qml.QNode, qml.qnode_old.QNode, qml.ExpvalCost)):
             # Input is a QNode:
             # result = some_transform(qnode, *transform_args)(*qnode_args)
             wrapper = self.qnode_wrapper(qnode, targs, tkwargs)
