+++ tests/test_measure.py
@@ -615,9 +615,7 @@ class TestState:
 
         with pytest.raises(
             qml.QuantumFunctionError,
-            match="The state or density matrix"
-            " cannot be returned in combination"
-            " with other return types",
+            match="Computing the gradient of circuits that return the state is not supported",
         ):
             func()
 
@@ -760,7 +758,7 @@ class TestState:
             qml.RY(x, wires=0)
             return state()
 
-        x = tf.Variable(0.1, dtype=tf.complex128)
+        x = tf.Variable(0.1, dtype=tf.float64)
 
         with tf.GradientTape() as tape:
             result = func(x)
