+++ tests/transforms/test_batch_params.py
@@ -51,7 +51,7 @@ def test_autograd(diff_method, tol):
     dev = qml.device("default.qubit", wires=2)
 
     @qml.batch_params
-    @qml.beta.qnode(dev, diff_method=diff_method)
+    @qml.qnode(dev, diff_method=diff_method)
     def circuit(x):
         qml.RX(x, wires=0)
         qml.RY(0.1, wires=1)
@@ -77,7 +77,7 @@ def test_jax(diff_method, tol):
     dev = qml.device("default.qubit", wires=2)
 
     @qml.batch_params
-    @qml.beta.qnode(dev, interface="jax", diff_method=diff_method)
+    @qml.qnode(dev, interface="jax", diff_method=diff_method)
     def circuit(x):
         qml.RX(x, wires=0)
         qml.RY(0.1, wires=1)
@@ -103,7 +103,7 @@ def test_jax_jit(diff_method, tol):
     dev = qml.device("default.qubit", wires=2)
 
     @qml.batch_params
-    @qml.beta.qnode(dev, interface="jax", diff_method=diff_method)
+    @qml.qnode(dev, interface="jax", diff_method=diff_method)
     def circuit(x):
         qml.RX(x, wires=0)
         qml.RY(0.1, wires=1)
@@ -129,7 +129,7 @@ def test_torch(diff_method, tol):
     dev = qml.device("default.qubit", wires=2)
 
     @qml.batch_params
-    @qml.beta.qnode(dev, interface="torch", diff_method=diff_method)
+    @qml.qnode(dev, interface="torch", diff_method=diff_method)
     def circuit(x):
         qml.RX(x, wires=0)
         qml.RY(0.1, wires=1)
@@ -157,7 +157,7 @@ def test_tf(diff_method, tol):
     dev = qml.device("default.qubit", wires=2)
 
     @qml.batch_params
-    @qml.beta.qnode(dev, interface="tf", diff_method=diff_method)
+    @qml.qnode(dev, interface="tf", diff_method=diff_method)
     def circuit(x):
         qml.RX(x, wires=0)
         qml.RY(0.1, wires=1)
@@ -184,7 +184,7 @@ def test_tf_autograph(tol):
     dev = qml.device("default.qubit", wires=2)
 
     @qml.batch_params
-    @qml.beta.qnode(dev, interface="tf", diff_method="backprop")
+    @qml.qnode(dev, interface="tf", diff_method="backprop")
     def circuit(x):
         qml.RX(x, wires=0)
         qml.RY(0.1, wires=1)
@@ -235,7 +235,7 @@ def test_unbatched_parameter():
     dev = qml.device("default.qubit", wires=1)
 
     @qml.batch_params
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def circuit(x, y):
         qml.RY(x, wires=[0])
         qml.RX(y, wires=[0])
@@ -255,7 +255,7 @@ def test_initial_unbatched_parameter():
     dev = qml.device("default.qubit", wires=1)
 
     @qml.batch_params
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def circuit(x, y):
         qml.RY(x, wires=[0])
         qml.RX(y, wires=[0])
