+++ tests/transforms/test_draw.py
@@ -30,7 +30,7 @@ def test_drawing():
 
     dev = qml.device("default.qubit", wires=2)
 
-    @qml.beta.qnode(dev, interface="autograd")
+    @qml.qnode(dev, interface="autograd")
     def circuit(p1, p2=y, **kwargs):
         qml.RX(p1, wires=0)
         qml.RY(p2[0] * p2[1], wires=1)
@@ -57,7 +57,7 @@ def test_drawing_tf():
 
     dev = qml.device("default.qubit", wires=2)
 
-    @qml.beta.qnode(dev, interface="tf")
+    @qml.qnode(dev, interface="tf")
     def circuit(p1, p2=y, **kwargs):
         qml.RX(p1, wires=0)
         qml.RY(p2[0] * p2[1], wires=1)
@@ -84,7 +84,7 @@ def test_drawing_torch():
 
     dev = qml.device("default.qubit", wires=2)
 
-    @qml.beta.qnode(dev, interface="torch")
+    @qml.qnode(dev, interface="torch")
     def circuit(p1, p2=y, **kwargs):
         qml.RX(p1, wires=0)
         qml.RY(p2[0] * p2[1], wires=1)
@@ -112,7 +112,7 @@ def test_drawing_jax():
 
     dev = qml.device("default.qubit", wires=2)
 
-    @qml.beta.qnode(dev, interface="jax")
+    @qml.qnode(dev, interface="jax")
     def circuit(p1, p2=y, **kwargs):
         qml.RX(p1, wires=0)
         qml.RY(p2[0] * p2[1], wires=1)
@@ -139,7 +139,7 @@ def test_drawing_ascii():
 
     dev = qml.device("default.qubit", wires=2)
 
-    @qml.beta.qnode(dev, interface="autograd")
+    @qml.qnode(dev, interface="autograd")
     def circuit(p1, p2=y, **kwargs):
         qml.RX(p1, wires=0)
         qml.RY(p2[0] * p2[1], wires=1)
@@ -162,7 +162,7 @@ def test_show_all_wires_error():
 
     dev = qml.device("default.qubit", wires=[-1, "a", "q2", 0])
 
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def circuit():
         qml.Hadamard(wires=-1)
         qml.CNOT(wires=[-1, "q2"])
@@ -178,7 +178,7 @@ def test_missing_wire():
 
     dev = qml.device("default.qubit", wires=["a", -1, "q2"])
 
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def circuit():
         qml.Hadamard(wires=-1)
         qml.CNOT(wires=["a", "q2"])
@@ -221,7 +221,7 @@ def test_invalid_wires():
     ordering does not exist on the device"""
     dev = qml.device("default.qubit", wires=["a", -1, "q2"])
 
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def circuit():
         qml.Hadamard(wires=-1)
         qml.CNOT(wires=["a", "q2"])
@@ -237,7 +237,7 @@ def test_draw_batch_transform():
     dev = qml.device("default.qubit", wires=1)
 
     @functools.partial(qml.gradients.param_shift, shift=0.2)
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def circuit(x):
         qml.Hadamard(wires=0)
         qml.RX(x, wires=0)
@@ -254,7 +254,7 @@ def test_direct_qnode_integration():
     """Test that a QNode renders correctly."""
     dev = qml.device("default.qubit", wires=2)
 
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def qfunc(a, w):
         qml.Hadamard(0)
         qml.CRX(a, wires=[0, 1])
@@ -280,7 +280,7 @@ def test_same_wire_multiple_measurements():
     """Test that drawing a QNode with multiple measurements on certain wires works correctly."""
     dev = qml.device("default.qubit", wires=4)
 
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def qnode(x, y):
         qml.RY(x, wires=0)
         qml.Hadamard(0)
@@ -306,7 +306,7 @@ def test_same_wire_multiple_measurements_many_obs():
     """
     dev = qml.device("default.qubit", wires=4)
 
-    @qml.beta.qnode(dev)
+    @qml.qnode(dev)
     def qnode(x, y):
         qml.RY(x, wires=0)
         qml.Hadamard(0)
@@ -332,7 +332,7 @@ class TestWireOrdering:
 
         dev = qml.device("default.qubit", wires=["a", -1, "q2"])
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def circuit():
             qml.Hadamard(wires=-1)
             qml.CNOT(wires=["a", "q2"])
@@ -353,7 +353,7 @@ class TestWireOrdering:
 
         dev = qml.device("default.qubit", wires=["a", -1, "q2"])
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def circuit():
             qml.Hadamard(wires=-1)
             qml.CNOT(wires=["a", "q2"])
@@ -374,7 +374,7 @@ class TestWireOrdering:
 
         dev = qml.device("default.qubit", wires=[-1, "a", "q2", 0])
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def circuit():
             qml.Hadamard(wires=-1)
             qml.CNOT(wires=[-1, "q2"])
@@ -396,7 +396,7 @@ class TestWireOrdering:
 
         dev = qml.device("default.qubit", wires=[-1, "a", "q2", 0])
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def circuit():
             qml.Hadamard(wires=-1)
             qml.CNOT(wires=[-1, "q2"])
@@ -411,7 +411,7 @@ class TestWireOrdering:
 
         dev = qml.device("default.qubit", wires=["a", -1, "q2"])
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def circuit():
             qml.Hadamard(wires=-1)
             qml.CNOT(wires=["a", "q2"])
@@ -453,7 +453,7 @@ class TestWireOrdering:
         ordering does not exist on the device"""
         dev = qml.device("default.qubit", wires=["a", -1, "q2"])
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def circuit():
             qml.Hadamard(wires=-1)
             qml.CNOT(wires=["a", "q2"])
@@ -467,7 +467,7 @@ class TestWireOrdering:
         """Test that a QNode with no operations still draws correctly"""
         dev = qml.device("default.qubit", wires=3)
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def qnode():
             return qml.expval(qml.PauliX(wires=[0]) @ qml.PauliX(wires=[1]) @ qml.PauliX(wires=[2]))
