+++ tests/transforms/test_metric_tensor.py
@@ -249,7 +249,7 @@ class TestMetricTensor:
         correctly when the QNode contains a template"""
         dev = qml.device("default.qubit", wires=3)
 
-        @qml.beta.qnode(dev, expansion_strategy=strategy)
+        @qml.qnode(dev, expansion_strategy=strategy)
         def circuit(weights):
             qml.templates.StronglyEntanglingLayers(weights, wires=[0, 1, 2])
             return qml.probs(wires=[0, 1])
